Fix project permissions
