+++ examples/organization/src/OrganizationPage.tsx
@@ -1,10 +1,10 @@
 import { ID } from "jazz-tools";
 import { useParams } from "react-router";
 import { Layout } from "./Layout.tsx";
+import { CreateProject } from "./components/CreateProject.tsx";
 import { Heading } from "./components/Heading.tsx";
-import { InviteLink } from "./components/organization/InviteLink.tsx";
-import { OrganizationMembers } from "./components/organization/OrganizationMembers.tsx";
-import { CreateProject } from "./components/project/CreateProject.tsx";
+import { InviteLink } from "./components/InviteLink.tsx";
+import { OrganizationMembers } from "./components/OrganizationMembers.tsx";
 import { useCoState } from "./main.tsx";
 import { Organization } from "./schema.ts";
 
@@ -60,7 +60,7 @@ export function OrganizationPage() {
               </p>
             )}
             <div className="p-4 sm:p-6">
-              <CreateProject organizationId={organization.id} />
+              <CreateProject organization={organization} />
             </div>
           </div>
         </div>
