+++ examples/organization/src/components/CreateOrganization.tsx
@@ -1,13 +1,9 @@
 import { Group, ID } from "jazz-tools";
 import { useState } from "react";
 import { useNavigate } from "react-router";
-import { useAccount, useCoState } from "../../main.tsx";
-import {
-  DraftOrganization,
-  ListOfProjects,
-  Organization,
-} from "../../schema.ts";
-import { Errors } from "../Errors.tsx";
+import { useAccount, useCoState } from "../main.tsx";
+import { DraftOrganization, ListOfProjects, Organization } from "../schema.ts";
+import { Errors } from "./Errors.tsx";
 import { OrganizationForm } from "./OrganizationForm.tsx";
 
 export function CreateOrganization() {
